Added OSX to Travis CI.
Added logic in test.sh for building and unit testing on OSX in travis-ci.

Skip linux tests on osx and vice versa, to keep CI fast and save travis-ci resources.

Note: travis-ci on osx does not support docker. Therefore, a docker alias is created to enable docker commands to execute successfully for osx build and test.